LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-stretchy-001-expected.txt


PASS Operator dictionary chunk 1 - stretchy
FAIL Operator dictionary chunk 2 - stretchy assert_approx_equals: Stretchy property for ⌢ postfix should be 'true' expected 100 +/- 1 but got 50
PASS Operator dictionary chunk 3 - stretchy
PASS Operator dictionary chunk 4 - stretchy
PASS Operator dictionary chunk 5 - stretchy
stretchy for "⌢" (postfix):
⌢
 VS
⌢
